requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
azure-communication-callautomation
azure-communication-identity
azure-cosmos

# Azure Bot Service dependencies
botbuilder-core>=4.15.0
botbuilder-schema>=4.15.0
botframework-connector>=4.15.0

# OpenAI dependencies for bot intelligence
openai>=1.12.0

# HTTP client library (required by bot framework and OpenAI)
aiohttp>=3.8.0
